pyproject.toml
[tool.poetry]
packages = [{include = 'hough_transform_project', from='src'}]
name = "hough_transform"
version = "0.1.0"
description = ""
authors = ["Vlad Averkov <averkov.ve@phystech.edu>"]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.0"
pytest-cov = "^5.0.0"
nox = "^2024.4.15"
numpy = "^1.26.4"
pillow = "^10.3.0"
desert = "^2022.9.22"
marshmallow = "^3.21.1"

[tool.poetry.dev-dependencies]
coverage = "^7.5.0"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
mypy = "^1.10.0"
pytype = {version = "^2024.4.11", python = "3.7"}
typeguard = "^4.2.1"
flake8-annotations = "^3.0.1"
flake8-docstrings = "^1.7.0"
darglint = "^1.8.1"
xdoctest = "^1.1.3"
Sphinx = "^7.3.7"
nox = "^2024.4.15"
click = "^8.1.7"
pytest = "^8.1.2"
sphinx-autodoc-typehints = "^2.1.0"
numpy = "^1.26.4"
pillow = "^10.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

description = "The Hough Transform"
license = "MIPT"
readme = "README.md"
homepage = "https://github.com/YablochniyBoss/Hough_Transform"
repository = "https://github.com/YablochniyBoss/Hough_Transform"
keywords = ["Hough Transform"]


[tool.poetry.scripts]
hough = "hough_transform_project.console:main"

[tool.coverage.report]
fail_under = 100
